Shifted to LongURL backend
I had checked out LongURL before the previous commit, but I viewed
it as a product which was not quite the same -- the aim of this
was to implement searches locally, and get a speed boost from this.

However, it turns out that many of the URL shortening services
do not provide easy access to their reverse lookup functionality.
My choice is between writing a requests-based system that visits
each page and using a service like LongURL which presumably has
already done this.

The benefit of LongURL is historic data. URLs that no longer
resolve can be transformed using their database. It would be
better if we could also do some sort of batch processing like
bit.ly allowed, but I'll work with what I have.

This now means that basically any shortened URLs are successfully
expanded.

As part of the shift, I rewrote the system so it is set up to
apply generic transforms to URLs. I also repurposed some of the
code from the bit.ly expander to write my first non-expanding
transform (shortening in bit.ly). However, I have not yet resolved
on how I should pass configuration details to this service --
perhaps I was hasty in removing the config-file-reader code.