CLI argument passing; transform_all()
I implemented CLI argument passing for the transforms, which at
the moment only impacts on the shorten transform.

I added transform_all() to iterate through transforms with the
appropriate delay, and rewrote transform_in_place to make use of
it. I added an option to override the rate-limiting.

I renamed the project and updated the README.