Created entitle transform
I've written a piece of code for the entitle transform which
may have to be inspected when I'm more cogent. The idea is that
it uses LongURL to get title information on shortened URLs.

However, I'll have to implement my own method for titling other
URLs, so this could turn out to be largely redundant.

Also in this commit, I re-added the main module! I had not
remembered to include it after renaming the file in a previous
commit.